+++ src/natcap/invest/utils.py
@@ -158,8 +158,8 @@ def prepare_workspace(
             logging.captureWarnings(True)
             # If invest is launched as a subprocess (e.g. the Workbench)
             # the parent process can rely on this announcement to know the
-            # logfile path, and to know the invest process has started.
-            LOGGER.log(100, 'Writing log messages to %s', logfile)
+            # logfile path (within []), and to know the invest process started.
+            LOGGER.log(100, 'Writing log messages to [%s]', logfile)
             start_time = time.time()
             try:
                 yield